jitai/requirements.txt
matplotlib==3.0.1
beautifulsoup4==4.6.3
nltk==3.3
numpy==1.15.3
pandas==0.23.4
scikit-learn==0.20.0
scipy==1.1.0
six==1.11.0
requests==2.20.0
lxml==4.2.5
jpholiday==0.0.6
PyYAML==3.13